Ready for production

+++ projekt/LibraryApi/Controllers/ReservationController.cs
@@ -103,6 +103,10 @@ public class ReservationController: ControllerBase
     public async Task<IActionResult> Delete(int id)
     {
         var reservation = await _context.Reservations.FindAsync(id);
+        if (reservation == null)
+        {
+            return NotFound();
+        }
         _context.Reservations.Remove(reservation);
         await _context.SaveChangesAsync();
         return NoContent();
